add response

+++ src/notes/notes.service.ts
@@ -1,15 +1,21 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { randomBytes } from 'crypto';
 import database from 'src/database';
 
 @Injectable()
 export class NotesService {
 
-    async deleteNotes(body): Promise<any> {
-        
-        const notes = await database.client.del(body.id)
+    async deleteNotes(body): Promise<HttpStatus> {
 
-        return notes
+        const notes = JSON.parse(await database.client.get(body.id));
+        
+        if(!notes){
+            throw new HttpException('Not found', HttpStatus.NOT_FOUND)
+        }
+        
+        await database.client.del(body.id)
+        throw new HttpException('OK', HttpStatus.OK)
+        
     }
 
     async updateNotes(dto): Promise<Message> {
@@ -17,8 +23,7 @@ export class NotesService {
         const notes = JSON.parse(await database.client.get(dto.id));
         
         if(!notes){
-            console.log('Заметка не найдена, чего собрался менять?')
-            return
+            throw new HttpException('Not found', HttpStatus.NOT_FOUND)
         }
 
         const note : Message = {
@@ -34,7 +39,7 @@ export class NotesService {
             await database.client.setEx(note.id, note.expires, JSON.stringify(note));
             return note
         }
-    
+        
         await database.client.set(note.id, JSON.stringify(note))
 
         return note 
@@ -52,11 +57,10 @@ export class NotesService {
             expires: dto.expires * 60 || null,
         }
 
-        if(dto.expires !== null ){
+        if(note.expires !== null ){
             await database.client.setEx(id, note.expires, JSON.stringify(note));
             return note
         }
-    
         await database.client.set(id, JSON.stringify(note))
     
         return note
@@ -67,7 +71,7 @@ export class NotesService {
         const keys = await database.client.keys('*');
         const values = await Promise.all(keys.map((key) => database.client.get(key)));
         const parsedData = values.map((item) => JSON.parse(item));
-
+        
         return parsedData;
       }
